requirements.txt
pandas==2.3.3
plotly==6.3.1
haversine==2.9.0
streamlit==1.50.0
pillow==11.3.0
folium==0.20.0
streamlit-folium==0.25.3